move docker log reader to docker subpkg
